+++ assets/package.json
@@ -32,6 +32,7 @@
   "dependencies": {
     "@absinthe/socket": "0.2.1",
     "@apollo/client": "3.10.3",
+    "@chakra-ui/react": "2.8.2",
     "@dagrejs/dagre": "1.0.4",
     "@docsearch/js": "3.5.2",
     "@docsearch/react": "3.5.2",
@@ -48,6 +49,7 @@
     "@nivo/tooltip": "0.83.0",
     "@pluralsh/design-system": "3.59",
     "@react-hooks-library/core": "0.6.0",
+    "@saas-ui/react": "2.8.8",
     "@tanstack/react-virtual": "3.0.1",
     "anser": "2.1.1",
     "apollo-absinthe-upload-link": "1.7.0",
@@ -63,6 +65,7 @@
     "escape-carriage": "1.3.1",
     "filesize": "10.0.12",
     "forge-core": "1.4.6",
+    "framer-motion": "11.3.19",
     "fuse.js": "6.6.2",
     "grommet": "2.34.2",
     "grommet-icons": "4.10.0",
