+++ assets/src/components/commandpalette/CommandPaletteShortcut.tsx
@@ -1,0 +1,13 @@
+import { useHotkeys } from '@saas-ui/react'
+
+import { Shortcut } from './shortcuts'
+
+export default function CommandPaletteShortcut({
+  shortcut,
+}: {
+  shortcut: Shortcut
+}) {
+  useHotkeys(shortcut.hotkeys, shortcut.action)
+
+  return undefined
+}
